+++ packages/jaspr_cli/lib/src/config.dart
@@ -1,5 +1,7 @@
-import 'dart:io';
+import 'dart:io' as io;
 
+import 'package:file/file.dart';
+import 'package:file/local.dart';
 import 'package:path/path.dart' as path;
 import 'package:yaml/yaml.dart';
 
@@ -8,9 +10,13 @@ import 'logging.dart';
 enum JasprMode { static, server, client }
 
 class Project {
-  Project(this.logger);
+  Project(this.logger, {FileSystem? fs, Never Function(int)? exitFn})
+    : _fs = fs ?? const LocalFileSystem(),
+      _exitFn = exitFn ?? io.exit;
 
   final Logger logger;
+  final FileSystem _fs;
+  final Never Function(int) _exitFn;
 
   YamlMap? get pubspecYaml => _pubspecYaml;
   YamlMap get requirePubspecYaml {
@@ -21,14 +27,14 @@ class Project {
         tag: Tag.cli,
         level: Level.critical,
       );
-      exit(1);
+      _exitFn(1);
     }
     return pubspecYaml;
   }
 
   late final YamlMap? _pubspecYaml = () {
     var pubspecPath = 'pubspec.yaml';
-    var pubspecFile = File(pubspecPath).absolute;
+    var pubspecFile = _fs.file(pubspecPath).absolute;
     if (!pubspecFile.existsSync()) {
       return null;
     }
@@ -37,7 +43,7 @@ class Project {
       return loadYaml(pubspecFile.readAsStringSync()) as YamlMap;
     } catch (e) {
       logger.write('Could not parse pubspec.yaml file: $e', tag: Tag.cli, level: Level.critical);
-      exit(1);
+      _exitFn(1);
     }
   }();
 
@@ -50,7 +56,7 @@ class Project {
         tag: Tag.cli,
         level: Level.critical,
       );
-      exit(1);
+      _exitFn(1);
     }
   }
 
@@ -71,7 +77,7 @@ class Project {
           defaultValue: true,
         );
         if (result) {
-          var result = Process.runSync('dart', ['pub', 'add', '--dev', 'jaspr_builder']);
+          var result = io.Process.runSync('dart', ['pub', 'add', '--dev', 'jaspr_builder']);
           if (result.exitCode != 0) {
             log.err(result.stderr as String?);
             logger.write(
@@ -79,7 +85,7 @@ class Project {
               tag: Tag.cli,
               level: Level.critical,
             );
-            exit(1);
+            _exitFn(1);
           }
 
           log.success('Successfully added jaspr_builder to your dev_dependencies.');
@@ -106,11 +112,11 @@ class Project {
     var configYaml = requirePubspecYaml['jaspr'];
     if (configYaml == null) {
       logger.write('Missing \'jaspr\' options in pubspec.yaml.', tag: Tag.cli, level: Level.critical);
-      exit(1);
+      _exitFn(1);
     }
     if (configYaml is! YamlMap) {
       logger.write('\'jaspr\' options must be a yaml map in pubspec.yaml.', tag: Tag.cli, level: Level.critical);
-      exit(1);
+      _exitFn(1);
     }
     return configYaml;
   }
@@ -134,42 +140,82 @@ class Project {
     var modeYaml = configYaml['mode'];
     if (modeYaml == null) {
       logger.write(
-        '\'jaspr.mode\' option is required but missing in pubspec.yaml.',
+        '\'jaspr.mode\' option in pubspec.yaml is required but missing.',
         tag: Tag.cli,
         level: Level.critical,
       );
-      exit(1);
+      _exitFn(1);
     }
     var modeOrNull = JasprMode.values.where((v) => v.name == modeYaml).firstOrNull;
     if (modeOrNull == null) {
       logger.write(
-        '\'jaspr.mode\' must be one of ${JasprMode.values.map((v) => v.name).join(', ')} in pubspec.yaml.',
+        '\'jaspr.mode\' in pubspec.yaml must be one of ${JasprMode.values.map((v) => v.name).join(', ')}.',
         tag: Tag.cli,
         level: Level.critical,
       );
-      exit(1);
+      _exitFn(1);
     }
     return modeOrNull;
   }
 
-  String? get devCommand {
+  List<String>? get target {
     var configYaml = _requireJasprOptions;
 
-    var devCommandYaml = configYaml['dev-command'];
-    if (devCommandYaml != null) {
-      if (devCommandYaml is! String) {
-        logger.write('\'jaspr.dev-command\' must be a string in pubspec.yaml.', tag: Tag.cli, level: Level.critical);
-        exit(1);
+    var targetYaml = configYaml['target'];
+    if (targetYaml != null) {
+      final targets = <String>[];
+      switch (targetYaml) {
+        case String s:
+          targets.add(s);
+        case List<Object?> l:
+          targets.addAll(l.cast<String>());
+        default:
+          logger.write(
+            '\'jaspr.target\' in pubspec.yaml must be a String or List<String>.',
+            tag: Tag.cli,
+            level: Level.critical,
+          );
+          _exitFn(1);
       }
 
-      return devCommandYaml;
+      for (var target in targets) {
+        if (!_fs.file(target).existsSync()) {
+          logger.write(
+            'The file "$target" specified by \'jaspr.target\' in pubspec.yaml does not exist.',
+            tag: Tag.cli,
+            level: Level.critical,
+          );
+          _exitFn(1);
+        }
+      }
+
+      return targets;
+    }
+    return null;
+  }
+
+  String? get port {
+    var configYaml = _requireJasprOptions;
+
+    var portYaml = configYaml['port'];
+    if (portYaml != null) {
+      if (portYaml is int) {
+        return portYaml.toString();
+      } else {
+        logger.write(
+          '\'jaspr.port\' in pubspec.yaml must be an integer.',
+          tag: Tag.cli,
+          level: Level.critical,
+        );
+        _exitFn(1);
+      }
     }
     return null;
   }
 
   late final YamlMap? pubspecLock = () {
     var pubspecLockPath = 'pubspec.lock';
-    var pubspecLockFile = File(pubspecLockPath).absolute;
+    var pubspecLockFile = _fs.file(pubspecLockPath).absolute;
 
     if (!pubspecLockFile.existsSync() && pubspecYaml?['resolution'] == 'workspace') {
       var n = 1;
@@ -178,7 +224,7 @@ class Project {
         if (parent == pubspecLockFile.path) {
           break;
         }
-        pubspecLockFile = File(path.join(parent, 'pubspec.lock'));
+        pubspecLockFile = _fs.file(path.join(parent, 'pubspec.lock'));
         if (pubspecLockFile.existsSync()) {
           break;
         }
@@ -191,7 +237,7 @@ class Project {
         return loadYaml(pubspecLockFile.readAsStringSync()) as YamlMap;
       } catch (e) {
         logger.write('Could not parse pubspec.lock file: $e', tag: Tag.cli, level: Level.critical);
-        exit(1);
+        _exitFn(1);
       }
     }
     return null;